refactor: utils.ts를 책임에 따라 url.ts, api.ts, render.ts로 분리
